[Documentation][Frontend] Documentation for plugins and small improvement (#1404)
**Context:** The documentation for MLIR plugins did not state anything
about `entry_points` nor relevant functions when loading MLIR plugins
from python. Also changed the type of the `pass_pipeline` kwarg. Instead
of it being a tuple of Passes, it is now a list of Passes. This is nicer
for developers.

**Description of the Change:** The `pass_pipeline` keyword argument is
now a list instead of a tuple. This list is converted into a tuple at
the moment the qnode primitive is bound.

Added documentation for MLIR plugins.

---------

Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>