Change the pattern rewrite infrastructure in quantum-to-ion lowering pass to use dialect conversion (#1442)
**Context:**

In this issue #1431, we found that crash messages that happen in
`quantum-to-ion` lowering pass are not preserved correctly in the
output. currently, some of the error messages are raised by an
assertion, however, we like to be able to gracefully exit in a way that
diagnostic messages are printed in a nice way.

One possible solution is to return a failure in the form of
`LogicalResult` and let the catalyst compiler driver handle the
diagnostic messages. However, with the current greedy pattern rewriting
infrastructure used, we do not have roll back procedure and a failure
would not stop the pass from continuing which causes trouble in the
subsequent pattern matchings in the case of a previous failure.

**Description of the Change:**

Instead of greedy pattern rewriting system, we now use a dialect
conversion system which supports roll backs and allow us to exit when a
failure is encountered. This then enables the compiler driver to print
the diagnostic messages nicely.

**Benefits:**

We exit gracefully with nice output error messages.
We use an infrastructure that is more suitable for a dialect converting
pass.

**Possible Drawbacks:**
Dialect converting infrastructure is generally slower that a greedy
rewrite due to the roll back housekeeping.

**Related GitHub Issues:**
#1431