Update dep version parsing (#1461)
**Context:** At the end of a release cycle, it would be convenient to
comment out the pennylane and lightning versions in the `.dep-versions`
file, as so:

```bash
jax=0.4.28
...

# For a custom PL version, update the package version here and at
# 'doc/requirements.txt
# pennylane=0.41.0-dev1

# For a custom LQ/LK version, update the package version here and at
# 'doc/requirements.txt'
# lightning=0.41.0-dev1
```

This allows keeping the correct format in place without explicitly
pinning the package version. However, comments are not currently
recognized when the `.dep-versions` file is parsed in `setup.py`.

**Description of the Change:** Replace the old file parsing, which
performed simple string splitting based on the number of characters
after `jax=`, `pennylane=` and `lightning=`, with more robust
regex-based parsing that also allows for lines to be commented out.

The pennylane and lightning versions introduced at the end of the
PennyLane/Lightning 0.40.0 release cycle have also been commented out.