[PLXPR] Add fixture to ensure capture has been disabled after test run. (#1445)
**Context:** When running with three or cores with pytest, it is
possible that the last three PLXPR tests are run in different processes.
The `test_counts` test will raise an error before calling
`qml.capture.disable()` which means that this particular worker will
keep this setting on. This doesn't happen with 2 or less runners because
the next two tests will re-enable and disable the capture mechanism.

**Description of the Change:** Skip test.

**Benefits:** Maybe the capture mechanism can be used with a context
manager and/or set with a fixture.

**Possible Drawbacks:**

**Related GitHub Issues:**