Ensure an `ion.yield` operation is returning a region output to all parallal protocol ops (#1475)
**Context:** During `quantum-to-ion` pass, when decomposing RX, RY and
MS gates to pulses in a `ParallelProtocolOp`'s region, the region was
terminated with yield ops without any operands. This means no SSA values
were actually yielded from the region. At the same time,
`ParallelProtocolOp` returns (variadic-length) values of type
`!quantum.bit`. This means the IR is ill-formed.

**Description of the Change:**
Ensure that all `ParallelProtocolOp`s will yield the input-ed qubit SSA
values by doing so in the builder, when building the region of the
`ParallelProtocolOp`.

**Benefits:** mlir is no longer ill-formed; increased readability