Fixed incorrect use of walrus operator (#1478)
Incorrect use of the walrus operator caused `argnums` to be assigned a
boolean instead of the value of rhs value.

**Context:**
Parameter `argnums` of `catalyst.grad()` and `catalyst.value_and_grad()`
was getting ignored. The gradients returned were of the first parameter
regardless of `argnums`.
In incorrect use of the walrus operator `:=` was found in the code.

**Description of the Change:**
Brackets were added to ensure correct assignment of value.

**Benefits:**
`argnums` now works as expected.

**Possible Drawbacks:**

**Related GitHub Issues:**
[#1477](https://github.com/PennyLaneAI/catalyst/issues/1477)

---------

Co-authored-by: David Ittah <dime10@users.noreply.github.com>
Co-authored-by: Rohan Nolan Lasrado <rohan-nolan.lasrado@hpe.com>