Create a runtime OQD device that generates OpenAPL (#1516)
**Context:**
After lowering the ion dialect to llvm dialect, and then llvm ir, with
calls to the runtime capi stubs (#1466), the final llvm ir needs to be
executed. The stubs will need to generate
[openapl](https://github.com/OpenQuantumDesign/openapl) in a JSON
schema, for example
https://github.com/OpenQuantumDesign/oqd-trical/blob/master/examples/7_example_sequential_parallels.json,
which is the common entry point for oqd backends.


**Description of the Change:**
- During ion dialect to llvm dialect lowering, the json spec for the ion
is passed in as a string to the device init operation's kwarg
- Fill the OQD capi stubs with openapl JSON schema generation.
- Update OQD device interface to call the newly defined OQD stubs.


**Benefits:**
Catalyst runtime can now generate programs that can be accepted by OQD
backends!

[sc-73710]

**This PR is inherited from #1479 , updating it to the new OQD file
organization brought by #1508 **

---------

Co-authored-by: Mehrdad Malek <39844030+mehrdad2m@users.noreply.github.com>