Improve wheel renaming in Build Nightly Release action (#1542)
In the "Prepare Wheels" job, which fixes up the platform string in the
wheel filename, search and replace the platform string for wheel
filenames matching both 'PennyLane_Catalyst-*' and
'pennylane_catalyst-*'. Then at the end, ensure that all wheel filenames
use the correct capitalization, 'PennyLane_Catalyst'.

This resolves an issue where uploading the wheels to TestPyPI failed
because the wheel names are now (for some reason we don't yet fulling
understand) all lowercase, and consequently the `rename` commands do not
match any files to fix up the platform name, hence the failure to upload
to TestPyPI.