Fixed a bug when autograph is used with a pass like merge-rotations (#1541)
**Context:**
issue #1501, found a bug where using autograph in conjunction with
catalyst passes causes a crash. This issue is introduced by PR #1349,
where the `apply_registered_pass_p` is removed and passes are appended
to the qnode `kwargs`. In case of using autograph, this `kwarg` calld
`pass-pipeline` is passed down to the quantum function which does not
support such `kwarg` hence, causing the crash.

**Description of the Change:**
This PR resolves the issue by poping the `pass-pipeline` element from
the `kwargs` before passing it to the quantum function, and instead,
pass it to the call function of the newly created `qnode`

**Benefits:**
autograph can work when having catalyst pass decorators.

**Possible Drawbacks:**
None

**Related GitHub Issues:**
[sc-84121]