Fix wheels file names in nightly build action (#1552)
This is a follow-up to #1542, which did not fully fix the wheels
file-naming patterns for the nightly-build action.

This change removes the step that replaced `pennylane_catalyst` with
`PennyLane_Catalyst` in the wheels file names after fixing up the
platform name string. Removing this step ensures that the wheels file
name and the files contained within the wheels follow the same naming
convention.

The wheels renaming step in the nightly-build action continues to
perform the `rename` command on wheels files with both the
`pennylane_catalyst-*` and `PennyLane_Catalyst-*` spellings. While we
now expect setuptools to always generate wheels with the lowercase
spelling, this should ensure the wheels get the correct platform string
should this change in the future.

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>