[WIP] Capture "cancel_inverses" transform (#1544)
**Context:** Catalyst should be able to support PL `cancel_inverses`,
since it has `passes.cancel_inverses`. So, Catalyst doesn’t need to use
PennyLane’s instructions for how to "expand" the `cancel_inverses`
transform:

```python

 @qjit(experimental_capture=True)
 def func(x: float):
     @qml.transforms.cancel_inverses
     @qml.qnode(dev)
     def circuit(x: float):
        qml.RX(x, wires=0)
        qml.Hadamard(wires=0)
        qml.Hadamard(wires=0)
        return qml.expval(qml.PauliZ(0))
```

**Description of the Change:** Register the PL
`cancel_inverses_transform` primitive, add the Catalyst
`remove-chained-self-inverse` pass to the pipeline, and evaluate the
plxpr for the inner jaxpr using the same interpreter.

**Benefits:** We don't have to decorate the original function with
Catalyst's `cancel_inverses`, just filling up the pass pipeline at
registration time and binding it where needed is sufficient.

**Drawbacks:** At the moment of writing, the transform and the QNode
must be embedded in a function in order for QJIT to be able to capture
the transform. This will change soon with @mudit2812 's work on this
issue.

[sc-84898]