Allow `qalloc_p` to lower to a dynamic number of wires (#1549)
**Context:**
Although the quantum.alloc operation in the mlir layer [allows both a
static and dynamic number of
wires](https://github.com/PennyLaneAI/catalyst/blob/main/mlir/include/Quantum/IR/QuantumOps.td#L117),
in the frontend when lowering `qalloc_p`, the static option is always
used. This means we cannot lower a circuit with a dynamic number of
wires to mlir.

**Description of the Change:**
Update `qalloc_p` lowering so it can handle a dynamic number of wires.
Correspondingly, when tracing, the dynamic jaxpr tracer is passed into
the alloc primitive when appropriate.

**Benefits:**
Catalyst can compile circuits with a dynamic number of wires.

**Possible Drawbacks:**
Measurements whose return shape depend on number of wires are still
static. This is because this PR is step 1 of splitting #1532 .

**Related GitHub Issues:**

[sc-84672]

---------

Co-authored-by: David Ittah <dime10@users.noreply.github.com>
Co-authored-by: Mehrdad Malek <39844030+mehrdad2m@users.noreply.github.com>