Allow `quantum.compbasis` op to take in quantum registers in mlir (#1553)
**Context:**
As part of the work to support dynamic number of wires in catalyst, the
`quantum.compbasis` operation should be able to take in a SSA value of a
quantum register type.

Currently a compbasis op can only take in an explicit list of qubits.
This creates problems, for example, when trying to return `probs()` or
`state()` from a circuit with a dynamic number of wires. Without the
explicit list of qubits known at compile time (we don't know how many
qubits there are), the compbasis value/op cannot be constructed.

**Description of the Change:**
`quantum.compbasis` op in the mlir layer can take in qregs.

**This PR does not introduce any frontend tracing changes. This PR is a
part of the prototype #1532. Further changes will be introduced in
future PRs, for the ease of review/workflow organization.**

**Benefits:**
One step closer to dynamic qubits.

**Related GitHub Issues:**

[sc-84672]