WIP fixes for custom tracing code in Catalyst
In particular, we frequently make assumptions on what the current trace
is, the type of tracers in the input, etc, which breaks composability
of JAX transforms. I don't know if it's worth the effort of going
through all the custom tracing code to see if those assumptions can
be removed.