Decompose gates if they are non differentiable and in grad context (#1562)
**Context:**
Currently, we decompose gates to a set of gates supported by the device
specified in the device toml file. If the final decomposed gates are
non-differentiable but used in a gradient context, They are not executed
because of the differentiability issue. However, decomposing these gates
into a set of differentiable gates, would unblock
grad method to execute the circuits.

**Description of the Change:**
added checks for being a non-differentiable gate in a gradient context
and if so decompose it further until we reach the stopping criteria.

**Benefits:**
Unblocks the ability to differentiate circuits containing
non-differentiable gates like `statePrep` or `Rot`

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-84810]

---------

Co-authored-by: Erick Ochoa Lopez <erick.ochoalopez@xanadu.ai>