Optimize verification/decomposition of differentiated ops based on activity (#1569)
This PR is a replacement for #1543, which does not employ dataflow-based
activity analysis but instead determines the activity of an operation
purely based on whether its parameters are constants or tracers. ~As a
further optimization, only floating point values are considered to be
able to carry activity.~ (I decided to remove this optimization to
reduce complexity.)

This PR is based on top of #1562.

[sc-84808]