Trace to compbasis op with qreg when measurements do not specify a set of wires (#1565)
**Context:**
Now that compbasis ops can take in qreg values (#1553) in mlir, we can
trace to that when we see a compbasis measurement without an explicit
wire list. This would include `state(), probs(), sample(), counts()`.

This is part 3 of the qubit-invariant compilation prototype in #1532.

**Description of the Change:**
The `AbstractObs` jax aval now has two "modes": the original qubit mode,
and the new qreg mode. They will lower to compbasis ops with qubits and
qreg in mlir, respectively.

When tracing, the number of wires on the device can be either static or
dynamic (allowed by #1549 ). In this PR we only deal with the static
case, for the ease of workflow organization and review.

The tracing logic for these parts look like the following:
```python
def trace_quantum_measurements():
    ...
    m_wires = o.wires if o.wires else range(len(device.wires))
    obs_tracers, nqubits = trace_observables(o.obs, qrp, m_wires)
    ...


def trace_observables(obs, ...):
    ...
    wires = obs.wires if (obs and len(obs.wires) > 0) else m_wires
    if obs is None:
        # not a named observable, so do compbasis_p.bind(...)
        ...
```

So the logic essentially says if observable has specified wires, use the
observable wires; else, use all device wires. The notion of "all device
wires" here is represented by a python `range` object from 0 up to the
number of wires (minus one). We adopt this notion, since we are only
dealing with a static number of wires in this PR. We also update this 
sentinel to not use a `range`, but simply `None`, to instruct the usage 
of all wires on the device.

Therefore, when we are about to bind a compbasis primitive, if we see
wires as `None`, we know we are taking measurement on all wires
of the device. In this case we can actualize (insert back all result
qubits) the qreg and enter the qreg compilation mode for the compbasis
op.

The `from_plxpr` pipeline is also accordingly updated.

**Benefits:**
The same compiled IR can be reused even if the device changes its number
of qubits across runs. We are missing a final link (bufferizing
measurement results based on the tracer on the `qml.device(wires=...)`),
but we are now very close!

**Possible Drawbacks:**
We lose verification on these measurement operations' return shapes when
measuring all wires in the device. We delay the verification to the
runtime. Although, one could make the argument that the verification is
not even necessary anymore in the measurement ops themselves, since we
now only have one source of truth (the result shape), as opposed to two
sources of truth (result shape and qubit list).

**Related GitHub Issues:**

[sc-86499]

---------

Co-authored-by: David Ittah <dime10@users.noreply.github.com>