Encapsulate Stim dependency in single library
Dependency to Stim is restricted to the QECUtils library. Classes from
that library are forward-declared in our header, and Stim headers are
only included in source files. This also requires that Stim classes
that are member variables in our classes be wrapped in a pointer(-like).

Some simplifications are made such as only relying on the
FlexPauliString class, instead of both it and the PauliString class.

Additionally, const correctness is improved for all QECUtils API
functions.