Capture and expand PL transforms (#1567)
**Context:** PL transforms without Catalyst pass counterpart should be
expanded by PL rules. When such PL transform is found, any embedded
transform that follows should be also expanded by PL rules, even if
there is Catalyst pass for it.

**Description of the Change:** Transforms are expanded using PL's
`ExpandTransformsInterpreter` , which is the interpreter behind the
magic of the `expand_plxpr_transforms()` method.

**Benefits:** Notice that with the usage of the mentioned interpreter,
we get for free the functionality that disables the conversion between
PL transforms and Catalyst passes once a non-supported transform is
found.

**Drawbacks:** Due to unsupported instruction `AmplitudeEmbedding` in
Catalyst, as a workaround `merge_amplitude_embedding` is followed by
'decompose'. It has to be decided if this is enough or we need to refine
the strategy. However, this is a task for a future PR.

[sc-86182]