Fix ASAN error for the PauliString class
Somehow copying FlexPauliStrings appears to lead to memory issues
(segfaults, heap buffer overflows) which manifested on when running
on Linux. ASAN also picked up on the issues, but doesn't flag on macOS.

Either way we stick to move semantics for the pauli string class and
our wrapper class, which seems to fix the issue and should provide
better performance.