Fix a bug when chained passes are used on a qnode (#1576)
**Context:**
In the current status of the Catalyst pass api, pass decorators recieve
a qnode and return a classical wrapper function. This make stacking
multiple passes/pipelines not possible. For more details please refer to
Issue #1547.

In the current status of the Catalyst pass api, there is significant
code duplication across different pass decorators (cancel_inverses,
merge_rotations, ions_decomposition, to_ppr, etc.) and the main pass API
functions (pipeline, apply_pass, apply_pass_plugin). Each of these
implements similar QNode wrapping logic, type checking, and pass
pipeline management. This duplication makes the code harder to maintain
and increases the risk of inconsistencies.

**Description of the Change:**
Added a new QNodeWrapper class which can be used to return in the
decorators.
Added a PassPipelineWrapper class that extends QNodeWrapper to handle
pass pipeline management
refactored the pass-api to eliminate duplicate codes. 

**Benefits:**
Catalyst's support for chained pass decorators is restored.
Reduced code duplicates.

**Possible Drawbacks:**
The new class hierarchy adds a layer of abstraction that might make the
code slightly harder to understand. Ideally we liked for the decorators
to receive and return a qnode. However, this is not possible at this
point.

**Related GitHub Issues:**
closes issue #1547 
Note that as a side benefit, this PR resolves issue #1548 
[sc-85468]
[sc-85470]