Update usage of while_loop to work with latest pl in `from_plxpr` (#1583)
**Context:**

In [PennyLane
#7084](https://github.com/PennyLaneAI/pennylane/pull/7084), I updated
`while_loop` to better support dynamic shapes, and removed
`abstract_shapes_slice` in the process. This is forcing a change to the
signature for `while_loop`.

**Description of the Change:**

Removes `abstract_shapes_slice` from the call signature to `while_loop`.

Adds a `NotImplementedError` if abstract shapes are passed to
`for_loop`. We can figure out how to implement this later. I added this
since I noticed that `abstract_shapes_slice` wasn't getting used, so I
thought it better to have a `NotImplementedError` then fall over in some
sort of way later.

**Benefits:**

Catalyst works with latest pennylane again.

**Possible Drawbacks:**

**Related GitHub Issues:**