Improve documentation and error messaging for .at method with multiple indices (#1595)
**Context:**
When using multiple indices with the `.at` method, the indices must be
sorted and unique. However, these requirements cannot be checked at
compile time, and one must explicitly provide these guarantees using the
parameters ``indices_are_sorted=True`` and ``unique_indices=True``. If
these parameters are not provided, the compiler will fail even if the
indices are unique and sorted.

**Description of the Change:**
Added this info to the jax integration page and also improived error
messaging to hint the user for the required `.at` parameters.

**Benefits:**
Better docs and error messaging

**Possible Drawbacks:**

**Related GitHub Issues:**
issue #1446 
[sc-81963]

---------

Co-authored-by: David Ittah <dime10@users.noreply.github.com>
Co-authored-by: Joey Carter <joseph.carter@xanadu.ai>