run split-multiple-tapes in a module (#1597)
**Context:** Split multiple-tapes creates new functions. This means it
can't be a function level pass, it must be a module level pass.

**Description of the Change:** Changes the test to run
split-multiple-tapes in a module.

It is unclear if this will fix the stochastic bug, but it is clear it is
a step in the right direction. I have some changes attempting to use
renameToUnique but I am a bit unhappy with them at the moment.