Fix autograph when using qml.ctrl/adjoint on an operator object (#1605)
**Context:**
when transforming functions, autograph handles some Catalyst wrapper
functions as special cases. This is added as a hotfix
[here](https://github.com/PennyLaneAI/catalyst/blob/845851662af490a45bff92ec32b5d1aaef7e7b10/frontend/catalyst/autograph/ag_primitives.py#L537-L552).
The hotfix assumes that catalyst wrapper functions always accept a
`callable`. PR #1212 Adds two `qml` functions (`qml.adjoint` and
`qml.ctrl`) to the list of specially handled wrappers. However, these
are not pure wrappers in the sense that their `arg[0]` can be an
operation object as opposed to a callable.

**Description of the Change:**
Handle cases where the argument to the specially handled functions are
operation objects and not callable.

TODO:
- [x] Add test for `qml.ctrl`
- [x] Add test for `qml.adjoint`
- [x] Add Changelog

**Benefits:**
autograph works in the case where `qml.ctrl` and `qml.adjoint` receive
an operation as argument.

**Possible Drawbacks:**

**Related GitHub Issues:**
issue #1339 
[sc-79375]