RC branch CI to test against other rc branches (!!!REVERT BEFORE RELEASE) (#1628)
**Context:**
Currently, CIs running on the Catalyst rc branch, test Catalyst with
respect to the Pennylane and Lightning versions specified in the
`.dep-versions` file. However, this can cause issues when there are
significant divergence between the main and rc branches in these repos.

**Description of the Change:**
In this PR we temporarily overwrite the installation of `Pennylane`,
`Pennylane-lightning`, and `Pennylane-Lightning-kokkos` to point to the
RC branches.

**Benefits:**
Better testing against other rc branches and avoid any conflicts with
dev branches of `Pennylane` and `Lightning`.

**Possible Drawbacks:**
This PR needs to be reverted before merging the rc branch to main.

**Related GitHub Issues:**