Add jaxpr constants as parameters when binding grad function. (#1613)
**Context:** The `jaxpr.consts` are not being resolved by JAX under some
circumstances. (It is unclear under which circumstances this happens at
the moment). However, `jaxpr.consts` are available during bind time.

**Description of the Change:** We pass `jaxpr.consts` as parameters to
the jax primitive. During stablehlo generation the jaxpr.consts are mlir
Values.

**Benefits:** The test passes.

**Possible Drawbacks:** This has the disadvantage of creating more
parameters as opposed to statically known values. E.g., consts are now
parameters as opposed to literals in the function itself. To mitigate
this, we only use the parameters as opposed to literals when the
literals are not made available by JAX.

**Related GitHub Issues:**

[sc-88455]

---------

Co-authored-by: Mehrdad Malek <39844030+mehrdad2m@users.noreply.github.com>