Add MBQC dialect (#1663)
**Context:** In order to QJIT compile a typical MBQC workload, we
require mechanisms to capture and lower the MBQC data structures and
functions recently added to the PennyLane
[`ftqc`](https://docs.pennylane.ai/en/stable/code/qml_ftqc.html) module.
In preparation for this work, we require an `mbqc` dialect in the MLIR
layer of Catalyst to contain the corresponding MBQC data types and
operations.

**Description of the Change:** This PR adds the framework for the `mbqc`
dialect. Currently, it only defines a single operation,
`mbqc.measure_in_basis`, which is an instruction representing
arbitrary-basis measurements parameterized by a plane along the major
axes of the Bloch sphere (either XY, ZX or YZ) and a rotation angle
about that plane. Other operations, data types and passes will be added
in future PRs as needed.


[[sc-89176](https://app.shortcut.com/xanaduai/story/89176/create-the-mbqc-mlir-dialect)]