Fix measurement handling in JAX tracer and device decomposition (#1634)
**Context:**
Fix issue when measurements handling in JAX on Qrack device were
incorrectly marked as modified when they weren’t.

**Description of the Change:**
- Removed validation for transformed quantum function outputs in
`is_transform_valid_for_batch_transforms`.
- Enhanced `apply_transform` to return `device_modify_measurements`
based on the presence and comparison of measurements in tapes.
- Updated `measurements_from_samples` to check for measurement types and
return a postprocessing function if no changes are detected. These
changes improve the handling of measurement results and streamline the
transformation process in the quantum function tracing.

**Benefits:**
- Fix bug

**Possible Drawbacks:**

**Related GitHub Issues:**
#1611

[[sc-88296]]

---------

Co-authored-by: erick-xanadu <110487834+erick-xanadu@users.noreply.github.com>
Co-authored-by: David Ittah <dime10@users.noreply.github.com>