Stop using AnyWires in prep for deprecation (#1667)
**Context:**

See PennyLane PR's https://github.com/PennyLaneAI/pennylane/pull/7312
and https://github.com/PennyLaneAI/pennylane/pull/7313 .

We want to stop using `WiresEnum` in favor of just having the default
`num_wires = None` to indicate that the number of wires should not be
validated.

**Description of the Change:**

Stop setting `num_wires` when it is `AllWires` and just depend on
inheriting the default.

**Benefits:**

Allows us to clean up the operator interface.

**Possible Drawbacks:**

**Related GitHub Issues:**