Fixing demo due to removal of `coeffs` from the `Sum` class (#1673)
This demo in Catalyst uses code that was just removed during
Deprecations and Removals for PL 0.42.

The reason why a warning message (indicating that the code was
deprecated) has not been caught before is that such a demo (and
potentially other ones) suppresses all warnings, including deprecation
warnings coming from PL.

I think this needs to change, for example, by suppressing all warnings
except for PL deprecation warnings, because a demo using deprecated PL
code should not be exposed to users.

[sc-89676]

---------

Co-authored-by: David Ittah <dime10@users.noreply.github.com>