Change logic around *when* standalone plugin is built (#1678)
**Context:**
Currently, the nightly build is failing due to the standalone plugin
being built and uploaded, where it shouldn't be built in the first
place.

Issue: When a github action is called using `workflow_call` the
`github.event_name` is not updated. This is something I was unaware of
prior to this.

A good question now would be, why not just align the arm64 workflow
if-statements to that of the other workflows, the expression being
`github.event.scheulde == '35 4 * * 4'`. The rationale explained in the
Benefits section.

**Description of the Change:**
This PR updates the logic around when we build the standalone plugins.

A new input is added to the four `build-wheel-` workflows. The input
being `build_standalone_plugin`, and it defaults to false.

Then we set an environment variable `BUILD_STANDALONE_PLUGIN` based on
the following logic:
- Does the input exist? 
- If yes, use the value of the input to determine if standalone plugin
build needed or not
    - If no, go to next
- Is the event_name 'schedule'?
    - If yes, build standalone plugin
    - If no, return false.

This environment variable is then used in the steps to determine if the
step should be run or not.

**Benefits:**
A minor benefit of this is that the steps themselves are more readable,
in that, `env.BUILD_STANDALONE_PLUGIN == 'true'` is easy to read and
know what the goal is.

The bigger, benefit is that it is best that we do not tie ourselves to
what the actual cron schedule is to build the standalone plugin. Though
I am sure it is a value that doesn't change too frequently, with the
current implementation, a change to the schedule would require
re-writing all the if-expressions, moreso, what if we wanted to start
building the standalone plugin twice a week? Or more?

The implementation proposed here moves us away from having to rely on
the value(s) of the cron schedule.


Another benefit is that we can opt-in to building standalone plugin when
we call these workflows from workflow_displatch or workflow_call.

**Possible Drawbacks:**
Possibility of bugs, but I have tested a build
[here](https://github.com/PennyLaneAI/catalyst/actions/runs/14619875471)
which finished without errors.

**Related GitHub Issues:**
[sc-89680]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>