Refactor `EnsureFunctionDeclaration` (#1683)
**Context:**
When lowering custom dialects to the llvm dialect, we will generate
calls to runtime CAPI functions. A utility function
`EnsureFunctionDeclaration()` generates declarations for the CAPI
functions in the llvm dialect IR.

Currently, this utility function is duplicated in the individual
dialect. We factor it out into a utility.

**Description of the Change:**
Factor out `EnsureFunctionDeclaration()` into the `Catalyst` dialect as
a utility.

**Benefits:**
Better code documentation; no duplicated code in multiple places.

**Related GitHub Issues:**
closes #1682 
[sc-89725]