Add pass for MBQC to LLVM dialect conversion (#1679)
**Context:** In #1663, we added the `mbqc` MLIR dialect. This dialect
currently contains one operation, `mbqc.measure_in_basis`, which
represents an instruction for arbitrary-basis measurements, where the
basis is parameterized by a plane in the Bloch sphere and an angle about
that plane. In order to QJIT compile an MBQC workload containing these
operations, we require a pass to convert the `mbqc.measure_in_basis`
operation into the LLVM dialect.

**Description of the Change:** This change adds the
`convert-mbqc-to-llvm` pass to perform the MBQC-to-LLVM dialect
conversion. This pass replaces the MLIR `mbqc.measure_in_basis`
operation with a call to the function
`__catalyst__mbqc__measure_in_basis()` in the IR. This change does
**not** add this stub to the Runtime C-API; this will be done in #1674.


[[sc-89606](https://app.shortcut.com/xanaduai/story/89606/create-pass-to-convert-operations-in-the-mbqc-dialect-to-the-llvm-dialect)]