Exit with success on skipped braket device pytests (#1698)
**Context:**
We recently skipped all the braket pytests. However, It was discovered
that `pytest.skip(reason="...", allow_module_level=True)` will not
collect the tests after skipping an entire pytest file, creating a
pytest exit code of 5 (which signals a failure).

**Description of the Change:**
We force a successful exit code after running the braket pytests in the
wheels scripts.

**Benefits:**
CI does not fail anymore.