Migrate `quantum` dialect to new one-shot bufferization   (#1686)
**Context:**
This work is based on #1027 

As part of the mlir update, the bufferization of the custom catalyst
dialects need to migrate to the new one-shot bufferization interface, as
opposed to the old pattern-rewrite style bufferization passes. See more
context in #1027.

As an example, here is how the new bufferization interface is used for
mlir's core `arith` dialect:

https://github.com/llvm/llvm-project/blob/main/mlir/lib/Dialect/Arith/Transforms/BufferizableOpInterfaceImpl.cpp

https://github.com/llvm/llvm-project/blob/main/mlir/include/mlir/Dialect/Arith/Transforms/BufferizableOpInterfaceImpl.h

https://github.com/llvm/llvm-project/blob/7ee0097b486b31be8b9a1750b2cd47580efd9587/mlir/lib/Dialect/Arith/IR/ArithDialect.cpp#L54


**Description of the Change:**
On the current mlir commit we track, both the old and new bufferization
styles exist. The old pattern rewrite style is deprecated.

To ease the workflow organization, we migrate one dialect at a time.
This PR migrates the `Quantum` dialect's bufferization to the new
one-shot interface.

Note that the new one-shot interface is supposed to be called only once
in the pipeline. However, because we haven't migrated all the dialects
yet, we simply swap out the old `--quantum--bufferize` pass in-place,
with the new one-shot bufferization pass running on the quantum dialect
only.


**Benefits:**
Align with mlir practices; one step closer to updating mlir.


[sc-71487]

---------

Co-authored-by: Tzung-Han Juang <tzunghan.juang@gmail.com>
Co-authored-by: paul0403 <paulhaochen.wang@gmail.com>
Co-authored-by: Paul <79805239+paul0403@users.noreply.github.com>
Co-authored-by: David Ittah <dime10@users.noreply.github.com>