[FTQC] Add support for parametric mid-circuit measurements (#1645)
**Context:**

PennyLane now contains an experimental
[`ftqc`](https://docs.pennylane.ai/en/stable/code/qml_ftqc.html) module
that defines a number of functions and abstractions for constructing
fault-tolerant quantum workloads in the MBQC representation. Among these
are the parametric, arbitrary-basis mid-circuit measurement operations,
such as
[`qml.ftqc.measure_arbitrary_basis()`](https://docs.pennylane.ai/en/stable/code/api/pennylane.ftqc.measure_arbitrary_basis.html).

To support these new functions and abstractions, we recently added the
`mbqc` dialect to Catalyst in #1663, which includes the
`mbqc.measure_in_basis` operation to represent the arbitrary-basis
mid-circuit measurements for MBQC workloads. We now require the
appropriate mechanisms to lower the arbitrary-basis MCM primitives
defined in PennyLane (added in
https://github.com/PennyLaneAI/pennylane/pull/7219) to the operations
defined in the `mbqc` MLIR dialect.

**Description of the Change:**

This PR registers the `measure_in_basis` PLxPR primitive in Catalyst and
adds the functions to lower it to the `mbqc.measure_in_basis` MLIR
operation. Recall that the basis state of an arbitrary-basis MCMs is
parameterized by a plane (one of XY, YZ or ZX) and a rotation angle
about that plane.

In order to execute the compiled workloads, we require a new pass
pipeline containing the MBQC-to-LLVM dialect-conversion pass. Currently,
this pipeline is defined manually in the tests, and consists of the
default quantum compilation pipeline with the `convert-mbqc-to-llvm`
pass inserted immediately before `llvm-dialect-lowering-pipeline`.

This PR also adds frontend lit tests and pytests to verify the generated
MLIR and to test the integration of the arbitrary-basis PLxPR
primitives, the MLIR operations and the corresponding Runtime C-API
call. The system-level pytests execute the compiled workloads on the
null.qubit device, and as such do not verify the correctness of the
results, only that we can properly integrate all of the components.

**Benefits:**

Support for arbitrary-basis measurements across the stack, which will
allow for compiling and executing full MBQC workloads.

**Possible Drawbacks:**

We do not currently have the ability to fully simulate compiled MBQC
workloads (e.g. on lightning.qubit). Rather, the aim of this PR is to
demonstrate that we can push an MBQC workload through the compilation
pipeline.


[[sc-88876](https://app.shortcut.com/xanaduai/story/88876/parametric-mid-circuit-measurements-are-supported-in-catalyst)]